requirements.txt
Flask==2.3.2
cryptography==41.0.3
pycryptodome==3.18.0